luffy/luffyapi/requirements.txt
Django==2.2.2
djangorestframework==3.11.0
Pillow==7.2.0
PyMySQL==0.9.3
pytz==2020.1
sqlparse==0.3.1
